updated file names and import strategy;
