--- src/assistant/__init__.py
@@ -1,1 +1,0 @@
-from .backend import ChatBot
