--- src/utils/__init__.py
@@ -1,2 +1,0 @@
-from .app_utils import create_directory
-from .ui_settings import UISettings
